add: notebook member valid

+++ src/components/NotebookSetting/NotebookMember.vue
@@ -95,7 +95,6 @@ import { Tag, Validation } from "@/models/vue-tags-input";
 import { Component, Prop, Vue, Watch } from "vue-property-decorator";
 import VueTagsInput from "@johmun/vue-tags-input";
 import RoleSelect from "@/components/NotebookSetting/RoleSelect.vue";
-import ro from "element-ui/src/locale/lang/ro";
 
 @Component({
   components: {
@@ -129,6 +128,50 @@ export default class NotebookMember extends Vue {
   @Watch("email")
   private watchEmail() {
     this.autocompleteEmail$.next(this.email);
+  }
+
+  private validRole(role: Role, member: MemberModel): boolean {
+    let result = false;
+    if (
+      role !== "owner" &&
+      member.id === this.$store.state.user.uid &&
+      this.members.filter(
+        value =>
+          value.status === "accept" &&
+          value.role === "owner" &&
+          value.id !== member.id
+      ).length === 0
+    ) {
+      this.$message.warning(this.$t("valid.memberRole") as string);
+    } else {
+      result = true;
+    }
+    return result;
+  }
+
+  private validDeleteMember(member: MemberModel): boolean {
+    let result = false;
+    if (
+      member.id === this.$store.state.user.uid &&
+      this.members.filter(
+        value => value.status === "accept" && value.id !== member.id
+      ).length === 0
+    ) {
+      this.$message.warning(this.$t("valid.memberDelete") as string);
+    } else if (
+      member.id === this.$store.state.user.uid &&
+      this.members.filter(
+        value =>
+          value.status === "accept" &&
+          value.role === "owner" &&
+          value.id !== member.id
+      ).length === 0
+    ) {
+      this.$message.warning(this.$t("valid.memberRole") as string);
+    } else {
+      result = true;
+    }
+    return result;
   }
 
   private getMembers() {
@@ -196,39 +239,43 @@ export default class NotebookMember extends Vue {
 
   private onDeleteMember(member: MemberModel) {
     log.debug("NotebookMember onDeleteMember", member);
-    this.$confirm(this.$t("confirm.deleteMember") as string, "Warning", {
-      confirmButtonText: this.$t("ok") as string,
-      cancelButtonText: this.$t("cancel") as string,
-      type: "warning"
-    })
-      .then(() => {
-        const loading = this.$loading({
-          lock: true,
-          text: this.$t("loading.deleting") as string
-        });
-        deleteMemberById(this.notebook.id, member.id)
-          .then(() => {
-            this.$message({
-              type: "success",
-              message: this.$t("deleted") as string
-            });
-            this.getMembers();
-          })
-          .catch(err => this.$message.error(err.message))
-          .finally(() => loading.close());
+    if (this.validDeleteMember(member)) {
+      this.$confirm(this.$t("confirm.deleteMember") as string, "Warning", {
+        confirmButtonText: this.$t("ok") as string,
+        cancelButtonText: this.$t("cancel") as string,
+        type: "warning"
       })
-      .catch(() => {});
+        .then(() => {
+          const loading = this.$loading({
+            lock: true,
+            text: this.$t("loading.deleting") as string
+          });
+          deleteMemberById(this.notebook.id, member.id)
+            .then(() => {
+              this.$message({
+                type: "success",
+                message: this.$t("deleted") as string
+              });
+              this.getMembers();
+            })
+            .catch(err => this.$message.error(err.message))
+            .finally(() => loading.close());
+        })
+        .catch(() => {});
+    }
   }
 
   private onChangeRole(role: Role, member: MemberModel) {
     log.debug("NotebookMember onChangeRole", role);
-    member.role = role;
-    memberRoleUpdate(this.notebook.id, member.id, role)
-      .then(() => {
-        this.notebook.roles[member.id] = role;
-        this.getMembers();
-      })
-      .catch(err => this.$message.error(err.message));
+    if (this.validRole(role, member)) {
+      member.role = role;
+      memberRoleUpdate(this.notebook.id, member.id, role)
+        .then(() => {
+          this.notebook.roles[member.id] = role;
+          this.getMembers();
+        })
+        .catch(err => this.$message.error(err.message));
+    }
   }
 
   private onAutocompleteEmail(keyword: string) {
